refactor: 스케쥴러 로직 서비스로 이동
